+++ runtime/rococo/src/weights/orml_tokens.rs
@@ -34,32 +34,32 @@ impl<T: frame_system::Config> orml_tokens::WeightInfo for WeightInfo<T> {
 	// Storage: System Account (r:1 w:1)
 	fn transfer() -> Weight {
 		Weight::from_ref_time(41_000_000)
-			.saturating_add(T::DbWeight::get().reads(4 as u64))
-			.saturating_add(T::DbWeight::get().writes(3 as u64))
+			.saturating_add(T::DbWeight::get().reads(4_u64))
+			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
 	// Storage: Tokens Accounts (r:2 w:2)
 	// Storage: AssetRegistry Assets (r:1 w:0)
 	// Storage: System Account (r:1 w:1)
 	fn transfer_all() -> Weight {
 		Weight::from_ref_time(44_000_000)
-			.saturating_add(T::DbWeight::get().reads(4 as u64))
-			.saturating_add(T::DbWeight::get().writes(3 as u64))
+			.saturating_add(T::DbWeight::get().reads(4_u64))
+			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
 	// Storage: Tokens Accounts (r:2 w:2)
 	// Storage: AssetRegistry Assets (r:1 w:0)
 	// Storage: System Account (r:1 w:1)
 	fn transfer_keep_alive() -> Weight {
 		Weight::from_ref_time(37_000_000)
-			.saturating_add(T::DbWeight::get().reads(4 as u64))
-			.saturating_add(T::DbWeight::get().writes(3 as u64))
+			.saturating_add(T::DbWeight::get().reads(4_u64))
+			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
 	// Storage: Tokens Accounts (r:2 w:2)
 	// Storage: AssetRegistry Assets (r:1 w:0)
 	// Storage: System Account (r:2 w:1)
 	fn force_transfer() -> Weight {
 		Weight::from_ref_time(41_000_000)
-			.saturating_add(T::DbWeight::get().reads(5 as u64))
-			.saturating_add(T::DbWeight::get().writes(3 as u64))
+			.saturating_add(T::DbWeight::get().reads(5_u64))
+			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
 	// Storage: Tokens Accounts (r:1 w:1)
 	// Storage: AssetRegistry Assets (r:1 w:0)
@@ -67,7 +67,7 @@ impl<T: frame_system::Config> orml_tokens::WeightInfo for WeightInfo<T> {
 	// Storage: System Account (r:1 w:1)
 	fn set_balance() -> Weight {
 		Weight::from_ref_time(29_000_000)
-			.saturating_add(T::DbWeight::get().reads(4 as u64))
-			.saturating_add(T::DbWeight::get().writes(3 as u64))
+			.saturating_add(T::DbWeight::get().reads(4_u64))
+			.saturating_add(T::DbWeight::get().writes(3_u64))
 	}
 }
